refactor!: move beditaClient to beditaApiClient
