+++ src/runtime/server/api/bedita/auth/change.patch.ts
@@ -1,6 +1,6 @@
 import { defineEventHandler, readBody } from 'h3';
 import { recaptchaVerifyToken } from '../../../utils/recaptcha';
-import { beditaClient, handleBeditaApiError } from '../../../utils/bedita-client';
+import { beditaApiClient, handleBeditaApiError } from '../../../utils/bedita-api-client';
 import { RecaptchaActions } from '../../../../utils/recaptcha-helpers';
 import type { UserAuth } from '../../../../types';
 import { filterUserDataToStore } from '../../../../utils/user-data-store';
@@ -10,7 +10,7 @@ export default defineEventHandler(async (event) => {
   try {
     const body = await readBody(event);
     await recaptchaVerifyToken(body?.recaptcha_token, RecaptchaActions.CHANGE_PASSWORD);
-    const client = await beditaClient(event);
+    const client = await beditaApiClient(event);
     const payload = {
       uuid: body?.uuid,
       password: body?.password,
