+++ src/runtime/server/api/bedita/auth/optout.post.ts
@@ -1,13 +1,13 @@
 import { defineEventHandler, readBody, setResponseStatus } from 'h3';
 import { recaptchaVerifyToken } from '../../../utils/recaptcha';
-import { beditaClient, handleBeditaApiError } from '../../../utils/bedita-client';
+import { beditaApiClient, handleBeditaApiError } from '../../../utils/bedita-api-client';
 import { RecaptchaActions } from '../../../../utils/recaptcha-helpers';
 
 export default defineEventHandler(async (event) => {
   try {
     const body = await readBody(event);
     await recaptchaVerifyToken(body.recaptcha_token, RecaptchaActions.OPTOUT);
-    const client = await beditaClient(event);
+    const client = await beditaApiClient(event);
     const response = await client.post('/auth/optout', {
       username: body.username,
       password: body.password,
