+++ src/runtime/server/api/bedita/auth/reset.post.ts
@@ -1,6 +1,6 @@
 import { defineEventHandler, readBody, getRequestURL, setResponseStatus } from 'h3';
 import { recaptchaVerifyToken } from '../../../utils/recaptcha';
-import { beditaClient, handleBeditaApiError } from '../../../utils/bedita-client';
+import { beditaApiClient, handleBeditaApiError } from '../../../utils/bedita-api-client';
 import { useRuntimeConfig } from '#imports';
 import { RecaptchaActions } from '../../../../utils/recaptcha-helpers';
 
@@ -10,7 +10,7 @@ export default defineEventHandler(async (event) => {
     const resetUrl = `${getRequestURL(event).origin}${runtimeConfig.bedita.resetPasswordPath}`;
     const body = await readBody(event);
     await recaptchaVerifyToken(body?.recaptcha_token, RecaptchaActions.RESET_PASSWORD);
-    const client = await beditaClient(event);
+    const client = await beditaApiClient(event);
     await client.post('/auth/change', {
       contact: body?.contact,
       change_url: resetUrl,
