+++ src/runtime/server/api/bedita/signup/signup.post.ts
@@ -1,6 +1,6 @@
 import { defineEventHandler, readBody } from 'h3';
 import { recaptchaVerifyToken } from '../../../utils/recaptcha';
-import { beditaClient, handleBeditaApiError } from '../../../utils/bedita-client';
+import { beditaApiClient, handleBeditaApiError } from '../../../utils/bedita-api-client';
 import { RecaptchaActions } from '../../../../utils/recaptcha-helpers';
 import { type JsonApiResourceObject } from '@atlasconsulting/bedita-sdk';
 
@@ -9,7 +9,7 @@ export default defineEventHandler(async (event) => {
     const body = await readBody(event);
 
     await recaptchaVerifyToken(body?.recaptcha_token, RecaptchaActions.SIGNUP);
-    const client = await beditaClient(event);
+    const client = await beditaApiClient(event);
     const response = await client.post('/signup', body);
 
     return response.data as JsonApiResourceObject;
